# Conflicts:
#	functions.php

+++ dt-mapping/geocode-api/mapbox-api.php
@@ -252,12 +252,14 @@ if ( ! class_exists( 'DT_Mapbox_API' ) ) {
 
         public static function load_mapbox_header_scripts() {
             // Mabox Mapping API
+            wp_enqueue_script( 'jquery-cookie', 'https://cdn.jsdelivr.net/npm/js-cookie@rc/dist/js.cookie.min.js', [ 'jquery' ], '3.0.0' );
+            wp_enqueue_script( 'mapbox-cookie', trailingslashit( get_stylesheet_directory_uri() ) . 'dt-mapping/geocode-api/mapbox-cookie.js', [ 'jquery', 'jquery-cookie' ], '3.0.0' );
             wp_enqueue_script( 'mapbox-gl', self::$mapbox_gl_js, [ 'jquery' ], self::$mapbox_gl_version, false );
             wp_enqueue_style( 'mapbox-gl-css', self::$mapbox_gl_css, [], self::$mapbox_gl_version );
         }
 
         public static function load_mapbox_search_widget() {
-            if ( file_exists( get_template_directory() . '/dt-mapping/geocode-api/mapbox.js' ) ) {
+            if ( file_exists( get_template_directory() . '/dt-mapping/geocode-api/mapbox-search-widget.js' ) ) {
                 global $post;
                 if ( is_single() ) {
                     $post_record = DT_Posts::get_post( get_post_type(), $post->ID );
@@ -266,12 +268,36 @@ if ( ! class_exists( 'DT_Mapbox_API' ) ) {
                 }
 
 
-                wp_enqueue_script( 'mapbox-search-widget', trailingslashit( get_stylesheet_directory_uri() ) . 'dt-mapping/geocode-api/mapbox.js', [ 'jquery', 'mapbox-gl', 'shared-functions' ], filemtime( get_template_directory() . '/dt-mapping/geocode-api/mapbox.js' ), true );
+                wp_enqueue_script( 'mapbox-search-widget', trailingslashit( get_stylesheet_directory_uri() ) . 'dt-mapping/geocode-api/mapbox-search-widget.js', [ 'jquery', 'mapbox-gl', 'shared-functions' ], filemtime( get_template_directory() . '/dt-mapping/geocode-api/mapbox-search-widget.js' ), true );
                 wp_localize_script(
                     "mapbox-search-widget", "dtMapbox", array(
                         'post_type' => get_post_type(),
                         "post_id" => $post->ID ?? 0,
                         "post" => $post_record ?? false,
+                        "map_key" => self::get_key(),
+                        "spinner_url" => get_stylesheet_directory_uri() . '/spinner.svg',
+                        "theme_uri" => get_stylesheet_directory_uri(),
+                        "translations" => array(
+                            'add' => __( 'add', 'disciple-tools' )
+                        )
+                    )
+                );
+                add_action( 'wp_head', [ 'DT_Mapbox_API', 'mapbox_search_widget_css' ] );
+            }
+        }
+
+        public static function load_mapbox_search_widget_users() {
+            if ( ! class_exists( 'Disciple_Tools_Users' ) ) {
+                return;
+            }
+            if ( file_exists( get_template_directory() . '/dt-mapping/geocode-api/mapbox-users-search-widget.js' ) ) {
+
+                wp_enqueue_script( 'mapbox-search-widget', trailingslashit( get_stylesheet_directory_uri() ) . 'dt-mapping/geocode-api/mapbox-users-search-widget.js', [ 'jquery', 'mapbox-gl', 'shared-functions' ], filemtime( get_template_directory() . '/dt-mapping/geocode-api/mapbox-users-search-widget.js' ), true );
+                wp_localize_script(
+                    "mapbox-search-widget", "dtMapbox", array(
+                        'post_type' => 'user',
+                        "user_id" => get_current_user_id(),
+                        "user_location" => Disciple_Tools_Users::get_user_location( get_current_user_id() ),
                         "map_key" => self::get_key(),
                         "spinner_url" => get_stylesheet_directory_uri() . '/spinner.svg',
                         "theme_uri" => get_stylesheet_directory_uri(),
@@ -361,14 +387,14 @@ if ( ! class_exists( 'DT_Mapbox_API' ) ) {
         public static function metabox_for_admin() {
             global $dt_mapping;
 
-            if ( isset( $_POST['mapbox_key'] )
-                 && ( isset( $_POST['geocoding_key_nonce'] )
+            if ( isset( $_POST['mapbox_key'] ) && isset( $_POST['action'] ) && ( isset( $_POST['geocoding_key_nonce'] )
                       && wp_verify_nonce( sanitize_text_field( wp_unslash( $_POST['geocoding_key_nonce'] ) ), 'geocoding_key' . get_current_user_id() ) ) ) {
 
                 $key = sanitize_text_field( wp_unslash( $_POST['mapbox_key'] ) );
-                if ( empty( $key ) ) {
+                $action = sanitize_text_field( wp_unslash( $_POST['action'] ) );
+                if ( empty( $key ) || $action === 'delete' ) {
                     self::delete_key();
-                } else {
+                } else if ( $action === 'add' ) {
                     self::update_key( $key );
                 }
             }
@@ -392,7 +418,12 @@ if ( ! class_exists( 'DT_Mapbox_API' ) ) {
                     <tr>
                         <td>
                             <?php wp_nonce_field( 'geocoding_key' . get_current_user_id(), 'geocoding_key_nonce' ); ?>
-                            Mapbox API Token: <input type="text" class="regular-text" name="mapbox_key" value="<?php echo ( $key ) ? esc_attr( $hidden_key ) : ''; ?>" /> <button type="submit" class="button">Update</button>
+                            Mapbox API Token: <input type="text" class="regular-text" name="mapbox_key" value="<?php echo ( $key ) ? esc_attr( $hidden_key ) : ''; ?>" />
+                            <?php if ( self::get_key() ) : ?>
+                                <button type="submit" name="action" value="delete" class="button">Delete</button>
+                            <?php else : ?>
+                                <button type="submit" name="action" value="add" class="button">Add</button>
+                            <?php endif; ?>
                         </td>
                     </tr>
                     <tr>
@@ -431,6 +462,10 @@ if ( ! class_exists( 'DT_Mapbox_API' ) ) {
                     </tbody>
                 </table>
             </form>
+            <style>
+                .connected{ padding: 10px; background-color: lightgreen;}
+                .not-connected { padding: 10px; background-color: lightcoral; }
+            </style>
             <br>
 
             <?php
